✨ feat: implement ComboBox component with FieldWrapper and integrate into AddVisitaDialog
