+++ src/app/(dashboard)/administrador/visitante/_components/combo-box/field-wrapper.tsx
@@ -1,5 +1,6 @@
+import { Label } from '@/components/ui/label'
 import { ReactNode } from 'react'
-import { Label } from '../label'
+
 
 interface FieldWrapperProps {
   label: string
@@ -13,7 +14,7 @@ export const FieldWrapper = ({
   className
 }: FieldWrapperProps) => {
   return (
-    <div className={`${className} flex  flex-col gap-2`}>
+    <div className={`${className} flex flex-col gap-2`}>
       <Label className="flex gap-2">{label}</Label>
       {children}
     </div>
