+++ src/app/(dashboard)/administrador/visitante/page.tsx
@@ -4,6 +4,7 @@ import { useVisitanteData } from '@/hooks/visitante/useVisitanteData'
 import { useState } from 'react'
 import { TableClient } from './_components/table-client'
 
+
 export default function PageVisitante() {
   const { data } = useVisitanteData()
   const [open, setOpen] = useState(false)
@@ -14,7 +15,7 @@ export default function PageVisitante() {
       <div className="flex w-full">
         <h2 className="text-2xl font-bold uppercase">Lista de Visitantes</h2>
       </div>
-      
+
       <TableClient data={data} open={open} setOpen={setOpen} />
     </section>
   )
