+++ src/components/ui/popover.tsx
@@ -46,4 +46,3 @@ function PopoverAnchor({
 }
 
 export { Popover, PopoverAnchor, PopoverContent, PopoverTrigger }
-
